refactor(rs-module): isolate legacy binary recovery in legacy/ dir
The recovery helpers exist solely to repair files written by the
v1.8-and-earlier web store path — bytes already on users' servers (or
already in their local IndexedDB cache) that don't render as valid JPEG/PNG
without intervention. They aren't library workarounds; they're
data-recovery shims with a definite end-of-life when no legacy data
remains.

- Move recoverLegacyBinaryStringEncoding to legacy/binary-recovery.ts.
- Extract the getFile string-branch into a sibling
  legacyBinaryStringToArrayBuffer helper, then collapse getFile's two
  branches into a single ternary.
- Re-export recoverLegacyBinaryStringEncoding from the package root so
  the existing web import keeps working.

No behavior change. All 188 web tests still pass.
