+++ packages/rs-module/src/index.ts
@@ -2,8 +2,13 @@ import { bookmarkSchema, noteSchema, imageMetaSchema, audioMetaSchema, videoMeta
 import type { InboxItem, InboxItemType, AppConfig, UserSettings, Collection, CollectionGroup } from './types.js';
 import type { MigrateResult } from 'rs-migrate';
 import { migrator, legacySchemas } from './migrations.js';
+import {
+  legacyBinaryStringToArrayBuffer,
+  recoverLegacyBinaryStringEncoding,
+} from './legacy/binary-recovery.js';
 export { migrator } from './migrations.js';
 export { wrapCodeBlock } from './migrations.js';
+export { recoverLegacyBinaryStringEncoding } from './legacy/binary-recovery.js';
 
 /** Current item types — legacy types like 'voice-memo' are excluded */
 const CURRENT_TYPES: Set<string> = new Set<string>([
@@ -43,57 +48,6 @@ function schemaAlias(type: string): string {
     case 'video': return 'video-meta';
     case 'document': return 'document-meta';
     default: return type;
-  }
-}
-
-/**
- * Recover the original bytes of a file that was uploaded by the legacy store
- * path (web app v1.8 and earlier).
- *
- * The legacy path converted the ArrayBuffer to a binary string with
- * `String.fromCharCode(byte)` per byte, then handed that string to
- * `storeFile`. The remotestoragejs sync layer then PUT the string body via
- * `fetch`, which UTF-8 encodes any code point > 0x7F — so a JPEG byte 0xFF
- * landed on the server as the two-byte sequence 0xC3 0xBF, doubling the size
- * of every high byte and making the file no longer a valid JPEG.
- *
- * The recovery is the exact inverse: UTF-8 decode the bytes back to the
- * binary string, then read each char's low byte to get the original byte.
- *
- * Detection is by structural test, not by per-format magic bytes:
- *
- *   - Real binary files (JPEG/PNG/etc.) almost always contain byte sequences
- *     that aren't valid UTF-8 (e.g. a lone 0xFF, which UTF-8 reserves as a
- *     never-valid lead byte). `TextDecoder({fatal:true})` throws on those,
- *     so we fall through and return the bytes unchanged.
- *   - Bytes that *do* decode as valid UTF-8 *and* whose every code point is
- *     < 256 can only have come from the binary-string-then-UTF-8 path —
- *     normal text with extended characters would produce code points >= 256
- *     for any non-Latin1 char, and arbitrary binary almost never round-trips
- *     through UTF-8 cleanly. So in practice this only fires on the legacy
- *     corruption pattern.
- *
- * Files that are already in the correct format pass through untouched
- * because of the throw on the first invalid byte. Returns the original
- * `ArrayBuffer` reference (not a copy) in that case.
- */
-export function recoverLegacyBinaryStringEncoding(buffer: ArrayBuffer): ArrayBuffer {
-  if (buffer.byteLength === 0) return buffer;
-  try {
-    const decoded = new TextDecoder('utf-8', { fatal: true }).decode(buffer);
-    for (let i = 0; i < decoded.length; i++) {
-      // If any code point is >= 256, this isn't a binary-string round-trip
-      // (the legacy encoder only ever produced chars 0–255). Bail out.
-      if (decoded.charCodeAt(i) >= 256) return buffer;
-    }
-    const recovered = new Uint8Array(decoded.length);
-    for (let i = 0; i < decoded.length; i++) {
-      recovered[i] = decoded.charCodeAt(i);
-    }
-    return recovered.buffer;
-  } catch {
-    // Invalid UTF-8 → already raw binary. Return as-is.
-    return buffer;
   }
 }
 
@@ -173,24 +127,13 @@ const InboxModule = {
           const file = await privateClient.getFile(path);
           if (!file?.data) return undefined;
           const mimeType = file.mimeType || file.contentType;
-          // remotestoragejs may return a binary string when the file was
-          // written by the legacy store path (`storeFile(mime, path,
-          // binaryString)`). One charCodeAt per char gives back the original
-          // bytes — that's the inverse of what the legacy encoder did.
-          if (typeof file.data === 'string') {
-            const bytes = new Uint8Array(file.data.length);
-            for (let i = 0; i < file.data.length; i++) {
-              bytes[i] = file.data.charCodeAt(i);
-            }
-            return { data: bytes.buffer, mimeType };
-          }
-          // ArrayBuffer path covers two cases:
-          //   - new uploads (correct raw bytes — passes through untouched)
-          //   - legacy uploads synced down from the server, where the sync
-          //     layer's UTF-8 corruption baked into the bytes on disk
-          // `recoverLegacyBinaryStringEncoding` handles the second case and
-          // is a no-op for the first.
-          return { data: recoverLegacyBinaryStringEncoding(file.data), mimeType };
+          // Both branches are no-ops for correctly-uploaded files; they
+          // recover bytes from the v1.8-and-earlier store path. See
+          // `legacy/binary-recovery.ts` for the detection invariant.
+          const data = typeof file.data === 'string'
+            ? legacyBinaryStringToArrayBuffer(file.data)
+            : recoverLegacyBinaryStringEncoding(file.data);
+          return { data, mimeType };
         },
 
         async getConfig(): Promise<AppConfig> {
